+++ sphinx/changelog.rst
@@ -7,8 +7,10 @@ dev version (forthcoming)
 ~~~~~~~~~~~~~~~~~~~~~~~~~
 
 - Add a `pos` keyword to filter paradigms on POS.
+- Usage of frequencies:
+    - Read frequencies from as much sources as possible in the Paralex package (Frequencies class).
+    - Weight cells based on the predictor-target pair frequency.
 - Prevent Matplotlib font manager from spamming the log in debug mode.
-- Adds a Frequency class able to handle multiple paralex sources for frequencies.
 - Switch patterns management to long format everywhere
 - Implement parallelisation for finding patterns and finding applicable patterns. See `cpus` config option.
 - Change management of sampling: `most_freq=False` replaced by `force_random=True`.
@@ -23,6 +25,7 @@ Version 2.0
 
 * Support for the Paralex standard.
 * Automatic generation of heatmaps for entropy computations.
+* Add a `cells` keyword to filter paradigms on cells.
 * Several bugfixes
 * Removal:
     * Support for wide paradigms.